added the authentication layer and successfully tested with swagger and postman
