+++ src/server/flask.py
@@ -1,0 +1,15 @@
+import os
+from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
+from dotenv import load_dotenv
+from src.src_context import root_dir
+load_dotenv(os.path.join(root_dir, '.env'))
+
+# creates Flask object
+app = Flask(__name__)
+app.config['SECRET_KEY'] = os.getenv('SECRET_KEY')
+app.config['SQLALCHEMY_DATABASE_URI'] = os.getenv('SQLALCHEMY_DATABASE_URI')
+app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = os.getenv("SQLALCHEMY_TRACK_MODIFICATIONS")
+db = SQLAlchemy(app)
+app.config['DEBUG'] = True
+
